requirements.txt
pytorch==1.10.0
tqdm==4.62.3
scikit-learn==0.22.1
tensorboardX==1.0